Refactor out CaseCourtReportContext from CaseCourtReport for easier testing of new field
